2009-05-16  Pavel Roskin  <proski@gnu.org>
	* script/lua/grub_lib.c (grub_lua_getenv): Make name and value
	const to avoid a warning.
	(grub_lua_setenv): Likewise.
	* script/lua/grub_main.c (grub_lua_parse_line): Use size_t for
	lmsg to fix a warning.

+++ script/lua/grub_main.c
@@ -51,7 +51,7 @@ grub_lua_parse_line (char *line, grub_reader_getline_t getline)
 
       if (r == LUA_ERRSYNTAX)
 	{
-	  int lmsg;
+	  size_t lmsg;
 	  const char *msg = lua_tolstring(state, -1, &lmsg);
 	  const char *tp = msg + lmsg - (sizeof(LUA_QL("<eof>")) - 1);
 	  if (grub_strstr(msg, LUA_QL("<eof>")) == tp)
